Defered a bug that causes opponent players to have the low hp highlight when above the threshold.
When the app sends a resync, players on full hp get the coloured background that indicates their hp is below the threshold set by the user. I need to test when this bug occurs and then fix it.
For now, I've increased the fullSyncPeriod to 10,000 changes which should reduce how often this problem occurs.